File Handling
The program reads strings from the user via the console until the user types "End". It stores these strings in an ArrayList, and then writes them out to both the console and a text file. Finally, the program identifies the highest string in the list according to lexicographical order and prints it out.